% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC-package.R
\docType{package}
\name{DataQC-package}
\alias{DataQC}
\alias{DataQC-package}
\title{DataQC: Basic Quality Control for Transcriptomics and Proteomics Data}
\description{
Provides tools for basic quality control (QC) of transcriptomics and proteomics data. This package includes functions to easily generate PCA plots, heatmaps, and boxplots, allowing users to perform initial data assessments and visualizations.
}
\author{
\strong{Maintainer}: Dong Zhang \email{785618641@qq.com} (\href{https://orcid.org/0000-0001-5551-9721}{ORCID})

}
\keyword{internal}
