% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_heatmap.R
\name{QC_heatmap}
\alias{QC_heatmap}
\title{Title}
\usage{
QC_heatmap(data, data_group = NULL, value_colour = NULL)
}
\arguments{
\item{data}{A data frame. The columns in the data box are the sample names and the rows are gene.
And the column names in the data must match row names in the data_group.}

\item{data_group}{A data frame. The row names in the data_group must match those in the data. The first column name must be id.
The second column name must be group.}

\item{value_colour}{The annotation_col must be a data frame.
The row names of annotation_col == the column names of data_heatmap.
The column names of annotation_col is the annotation legend name.}
}
\value{
A ggplot object.
}
\description{
Title
}
\examples{
library(dplyr)
library(DataQC)
data("iris")
data <- iris \%>\%
  dplyr::rename(group = "Species") \%>\%
  dplyr::select(-group)
data <- as.data.frame(t(data))
colnames(data) <- seq(1:150)
data_group <- iris \%>\%
 dplyr::rename(group = "Species") \%>\%
 dplyr::select(group)
data_group$id <- rownames(data_group)
data_group <- subset(data_group,select = c(id,group))
value_colour <- c("setosa" = "#00A087FF",# control group
                  "versicolor" = "#E64B35FF",# Experimental group
                  "virginica" = "#4DBBD5FF",# other group1
                  "other group2" = "#3C5488FF")# other group2
 QC_heatmap(data,data_group,value_colour)
}
